requirements.txt
Flask
gTTS
python-dotenv
google-generativeai
python-dotenv
